requirements.txt
streamlit
pandas
numpy
sentence-transformers
umap-learn
plotly
scikit-learn
